Add Invidious's download endpoint (#41)
* Add Invidious's download endpoint

* Fix lint

* Handle subtitle requests in download endpoint

* Fix download endpoint type logic

* Reword error messages

* Set Content-Deposition in /videoplayback as needed

* Fix issues with content-disposition encoding

* Fix boolean simplification

Co-authored-by: Alex Maras <dev@alexmaras.com>

* Use toString() rather than manual type cast

Co-authored-by: Alex Maras <dev@alexmaras.com>

* Use zod to validate downloadWidgetData

* Allow `ext` to be passed to label union object

Invidious also passes the ext attribute in the download_widget json
when downloading captions rather than only being present for media.

The only difference is that it remains unused due to the content-depo
header not being added for subtitle requests. This is an Invidious
bug.

* Remove redundant code

* Use regular zod union instead of discriminated

Discriminated unions requires z.literal or z.enum

* Typo

---------

Co-authored-by: Alex Maras <dev@alexmaras.com>